refactor(agent-daemon): extract shared adapter turn-lifecycle stamping
adapterSnapshotForTurnEvent is provider-agnostic (it maps a turn.* event
onto the lifecycle snapshot ADR 0008 requires at that transition); move it
with a generic stamping loop into turn_lifecycle_stamp.go so Phase B
adapters reuse it instead of duplicating the codex copy.
